EY-3638: Flytter kode for validering
