+++ libs/etterlatte-tilbakekreving-model/src/main/kotlin/TilbakekrevingBehandling.kt
@@ -1,6 +1,5 @@
 package no.nav.etterlatte.libs.common.tilbakekreving
 
-import no.nav.etterlatte.libs.common.feilhaandtering.ForespoerselException
 import no.nav.etterlatte.libs.common.sak.Sak
 import no.nav.etterlatte.libs.common.tidspunkt.Tidspunkt
 import java.util.UUID
@@ -29,14 +28,6 @@ data class TilbakekrevingBehandling(
             -> true
             else -> false
         }
-
-    fun validerGyldigTilbakekreving() {
-        val manglendeFelterVurdering = tilbakekreving.vurdering?.valider() ?: throw ManglerTilbakekrevingsvurdering()
-        val manglendeFelterPerioder = tilbakekreving.perioder.flatMap { periode -> periode.valider() }
-        if (manglendeFelterVurdering.isNotEmpty() || manglendeFelterPerioder.isNotEmpty()) {
-            throw UgyldigeFelterForTilbakekrevingsvurdering(manglendeFelterVurdering, manglendeFelterPerioder)
-        }
-    }
 
     companion object {
         fun ny(
@@ -69,23 +60,3 @@ enum class TilbakekrevingStatus {
         return this in listOf(OPPRETTET, UNDER_ARBEID, VALIDERT, UNDERKJENT)
     }
 }
-
-class UgyldigeFelterForTilbakekrevingsvurdering(
-    ugyldigeFelterVurdering: List<String>,
-    ugyldigeFelterPerioder: List<String>,
-) : ForespoerselException(
-        status = 400,
-        code = "UGYLDIGE_FELTER",
-        detail = "Et eller flere felter mangler gyldig verdi",
-        meta =
-            mapOf(
-                "ugyldigeFelterVurdering" to ugyldigeFelterVurdering,
-                "ugyldigeFelterPerioder" to ugyldigeFelterPerioder,
-            ),
-    )
-
-class ManglerTilbakekrevingsvurdering : ForespoerselException(
-    status = 400,
-    code = "MANGLER_VURDERING",
-    detail = "Tilbakekrevingen mangler vurdering",
-)
